CLDR-11828 integrate into XML
X-SVN-Rev: 14766

+++ tools/java/org/unicode/cldr/tool/RemoveEmptyCLDR.java
@@ -3,6 +3,7 @@ package org.unicode.cldr.tool;
 import java.io.File;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Map.Entry;
@@ -17,34 +18,47 @@ import com.google.common.collect.BiMap;
 import com.google.common.collect.HashBiMap;
 
 public class RemoveEmptyCLDR {
+
+    static final boolean PREFLIGHT = false;
+
     public static void main(String[] args) throws IOException {
+
         if (args.length == 0) {
             args = new String[] { "annotations", "annotationsDerived" };
         }
         Set<String> nonEmpty = new HashSet<>();
         BiMap<String, File> toDelete = HashBiMap.create();
+        int counter = 0;
         // eg /Users/markdavis/Google Drive/workspace/Generated/vxml/common/annotations
-        for (String dir : args) {
-            main: for (File f : new File(CLDRPaths.COMMON_DIRECTORY + dir).listFiles()) {
-                List<Pair<String, String>> data = new ArrayList<>();
-                String canonicalPath = f.getCanonicalPath();
-                if (!canonicalPath.endsWith(".xml") || canonicalPath.endsWith("root.xml")) {
+        for (String dirCommonSeed : Arrays.asList(CLDRPaths.SEED_DIRECTORY1, CLDRPaths.COMMON_DIRECTORY)) {
+            System.out.println("Checking: " + dirCommonSeed);
+            for (String dir : args) {
+                File dirFile = new File(dirCommonSeed + dir);
+                if (!dirFile.exists()) {
                     continue;
                 }
-                String name = f.getName();
-                name = name.substring(0,name.length()-4); // remove .xml
-                XMLFileReader.loadPathValues(canonicalPath, data, false);
-                for (Pair<String, String> item : data) {
-                    if (item.getFirst().contains("/identity")) {
+                main: for (File f : dirFile.listFiles()) {
+                    List<Pair<String, String>> data = new ArrayList<>();
+                    String canonicalPath = f.getCanonicalPath();
+                    if (!canonicalPath.endsWith(".xml") || canonicalPath.endsWith("root.xml")) {
                         continue;
                     }
-                    System.out.println("Skipping: " + canonicalPath);
-                    addNameAndParents(nonEmpty, name);
-                    continue main;
+                    String name = f.getName();
+                    name = name.substring(0,name.length()-4); // remove .xml
+                    XMLFileReader.loadPathValues(canonicalPath, data, false);
+                    for (Pair<String, String> item : data) {
+                        if (item.getFirst().contains("/identity")) {
+                            continue;
+                        }
+                        System.out.println(++counter + ") NOT-EMPTY: " + canonicalPath);
+                        addNameAndParents(nonEmpty, name);
+                        continue main;
+                    }
+                    toDelete.put(name, f);
                 }
-                toDelete.put(name, f);
             }
         }
+        counter = 0;
         // keep empty files that are needed for inheritance
         for (Entry<String, File> entry : toDelete.entrySet()) {
             String name = entry.getKey();
@@ -52,8 +66,10 @@ public class RemoveEmptyCLDR {
                 continue;
             }
             File file = entry.getValue();
-            System.out.println("Deleting: " + file.getCanonicalPath());
-            file.delete();
+            System.out.println(++counter + ") Deleting: " + file.getCanonicalPath());
+            if (!PREFLIGHT) {
+                file.delete();
+            }
         }
     }
